signal should be double

+++ src/LoadEventWorkspace2.h
@@ -15,7 +15,7 @@ public:
 
   void updateBoxType(std::vector<unsigned char> &boxType) const;
   void updateExtents(Eigen::Matrix<float, Eigen::Dynamic, 6> &extents) const;
-  void updateSignal(std::vector<float> &signal) const;
+  void updateSignal(std::vector<double> &signal) const;
   void updateEventIndex(Eigen::Matrix<uint64_t, Eigen::Dynamic, 2> &eventIndex) const;
 
 private:
